requirements.txt
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.1
httplib2==0.31.2
idna==3.11
lxml==6.0.2
mutagen==1.47.0
oauthlib==3.3.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pycryptodomex==3.23.0
pyparsing==3.3.2
python-docx==1.2.0
python-dotenv==1.2.2
requests==2.32.5
requests-oauthlib==2.0.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
websockets==16.0
yt-dlp==2026.3.17
yt-dlp-ejs==0.8.0
